--- .github/CODEOWNERS
@@ -1,1 +1,0 @@
-*	@auth0/sdk-team-approvers
