+++ .gitignore
@@ -10,6 +10,8 @@ lib-cov
 *.pid
 *.gz
 
+*.iml
+
 pids
 logs
 results
