--- bower.json
@@ -1,20 +1,0 @@
-{
-  "name": "jwt-decode",
-  "description": "Decode JWT payload in the browser",
-  "version": "2.2.0",
-  "keywords": ["jwt", "jsonwebtoken", "token", "decode"],
-  "homepage": "https://github.com/auth0/jwt-decode",
-  "main": "build/jwt-decode.js",
-  "dependencies": {},
-  "devDependencies": {},
-  "license": ["MIT"],
-  "ignore": [
-    "*/.*",
-    "*.json",
-    "src",
-    "*.yml",
-    "Gemfile",
-    "Gemfile.lock",
-    "*.md"
-  ]
-}
