+++ lib/atob.mjs
@@ -2,24 +2,16 @@
  * The code was extracted from:
  * https://github.com/davidchambers/Base64.js
  */
-
-var chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=';
-
-function InvalidCharacterError(message) {
-  this.message = message;
-}
-
-InvalidCharacterError.prototype = new Error();
-InvalidCharacterError.prototype.name = 'InvalidCharacterError';
-
-function polyfill (input) {
-  var str = String(input).replace(/=+$/, '');
+import { InvalidCharacterError } from "./JWTErrors.mjs"
+const chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=";
+function atob(input) {
+  const str = String(input).replace(/=+$/, "");
+  let output = "";
   if (str.length % 4 == 1) {
     throw new InvalidCharacterError("'atob' failed: The string to be decoded is not correctly encoded.");
   }
   for (
-    // initialize result and counters
-    var bc = 0, bs, buffer, idx = 0, output = '';
+    let bc = 0, bs, buffer, idx = 0;
     // get next character
     buffer = str.charAt(idx++);
     // character found in table? initialize bit storage and add its ascii value;
@@ -34,5 +26,4 @@ function polyfill (input) {
   return output;
 }
 
-
-module.exports = typeof window !== 'undefined' && window.atob && window.atob.bind(window) || polyfill;
+export { atob }
