+++ lib/base64_url_decode.mjs
@@ -1,4 +1,4 @@
-var atob = require('./atob');
+import { atob } from "./atob.mjs";
 
 function b64DecodeUnicode(str) {
   return decodeURIComponent(atob(str).replace(/(.)/g, function (m, p) {
@@ -10,7 +10,7 @@ function b64DecodeUnicode(str) {
   }));
 }
 
-module.exports = function(str) {
+export default function(str) {
   var output = str.replace(/-/g, "+").replace(/_/g, "/");
   switch (output.length % 4) {
     case 0:
