+++ lib/index.mjs
@@ -1,19 +1,13 @@
 'use strict';
 
-var base64_url_decode = require('./base64_url_decode');
+import base64_url_decode from "./base64_url_decode.mjs";
 
-function InvalidTokenError(message) {
-  this.message = message;
-}
+import { InvalidTokenError } from "./JWTErrors.mjs"
 
-InvalidTokenError.prototype = new Error();
-InvalidTokenError.prototype.name = 'InvalidTokenError';
-
-module.exports = function (token,options) {
+const decode = (token,options) => {
   if (typeof token !== 'string') {
     throw new InvalidTokenError('Invalid token specified');
   }
-
   options = options || {};
   var pos = options.header === true ? 0 : 1;
   try {
@@ -23,4 +17,4 @@ module.exports = function (token,options) {
   }
 };
 
-module.exports.InvalidTokenError = InvalidTokenError;
+export { decode, InvalidTokenError }
