+++ package.json
@@ -1,37 +1,17 @@
 {
-  "name": "jwt-decode",
-  "version": "2.2.0",
-  "description": "Decode JWT tokens, mostly useful for browser applications.",
-  "main": "lib/index.js",
+  "name": "@jrg/jwt-decode",
+  "version": "0.0.6",
+  "description": "Decode JWT tokens, mostly useful for browser applications. Based on https://github.com/auth0/jwt-decode",
+  "main": "lib/index.mjs",
   "keywords": [
     "jwt",
     "browser"
   ],
-  "scripts": {
-    "test": "grunt test"
-  },
   "repository": {
     "type": "git",
     "url": "git://github.com/auth0/jwt-decode"
   },
-  "author": "Jose F. Romaniello <jfromaniello@gmail.com> (http://joseoncode.com)",
+  "author": "Jackie Gleason",
   "license": "MIT",
-  "dependencies": {},
-  "devDependencies": {
-    "expect.js": "~0.2.0",
-    "grunt": "~0.4.0",
-    "grunt-browserify": "~5.0.0",
-    "grunt-cli": "~0.1.9",
-    "grunt-contrib-clean": "~0.5.0",
-    "grunt-contrib-connect": "~0.5.0",
-    "grunt-contrib-copy": "~0.4.1",
-    "grunt-contrib-uglify": "~0.2.4",
-    "grunt-contrib-watch": "~0.2.0",
-    "grunt-exec": "~0.4.2",
-    "grunt-s3": "~0.2.0-alpha.3",
-    "mocha": "~1.13.0",
-    "rimraf": "~2.2.2",
-    "testem": "~0.5.8",
-    "uglify-js": "^2.8.29"
-  }
+  "dependencies": {}
 }
