--- test_harness.html
@@ -1,29 +1,0 @@
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Mocha Tests Harness</title>
-  <link rel="stylesheet" href="/node_modules/mocha/mocha.css" />
-</head>
-<body>
-  <div id="mocha"></div>
-
-  <script src="/node_modules/mocha/mocha.js"></script>
-  <script src="/testem.js"></script>
-  <script src="/node_modules/expect.js/expect.js"></script>
-
-  <script src="/build/jwt-decode.js"></script>
-
-  <script>
-    mocha.timeout(60000);
-    mocha.ui('bdd');
-    mocha.reporter('html');
-    mocha.setup({ignoreLeaks: true});
-  </script>
-
-  <script src="/test/tests.js"></script>
-
-  <script>
-    mocha.run();
-  </script>
-</body>
-</html>
