--- testem.yml
@@ -1,46 +1,0 @@
-test_page: test_harness.html
-
-on_start:
-  command: /usr/bin/java -jar $HOME/.browserstack/BrowserStackTunnel.jar $BROWSERSTACK_KEY localhost,<port>,0 -force
-  wait_for_text: Press Ctrl-C to exit
-  wait_for_text_timeout: 9000
-
-launchers:
-
-  bs_firefox:
-    command: browserstack launch --timeout 120 --attach firefox:22.0 <url>/test_harness.html
-    protocol: browser
-
-  bs_chrome:
-    command: browserstack launch --timeout 120 --attach chrome:28.0 <url>/test_harness.html
-    protocol: browser
-
-  bs_safari_51:
-    command: browserstack launch --timeout 120 --attach safari:5.1 <url>/test_harness.html
-    protocol: browser
-
-  bs_ie_8:
-    command: browserstack launch --timeout 120 --attach ie:8.0 <url>/test_harness.html
-    protocol: browser
-
-  bs_ie_9:
-    command: browserstack launch --timeout 120 --attach ie:9.0 <url>/test_harness.html
-    protocol: browser
-
-  bs_ie_10:
-    command: browserstack launch --timeout 120 --attach ie:10.0 <url>/test_harness.html
-    protocol: browser
-
-  bs_android_23:
-    command: browserstack launch --timeout 240 --attach "Motorola Photon 4G" <url>/test_harness.html
-    protocol: browser
-
-  bs_android_41:
-    command: browserstack launch --timeout 240 --attach "Samsung Galaxy S III" <url>/test_harness.html
-    protocol: browser
-
-  bs_iphone_5:
-    command: browserstack launch --timeout 60 --attach "iPhone 5" <url>/test_harness.html
-    protocol: browser
-
-launch_in_ci: [bs_chrome, bs_firefox, bs_ie_8, bs_ie_9, bs_ie_10, bs_iphone_5, bs_android_41]
