--- testem_dev.yml
@@ -1,1 +1,0 @@
-test_page: test_harness.html
